Ft database tests

+++ models/entries.go
@@ -6,7 +6,7 @@ import (
 
 type Entry struct{
 	gorm.Model
-	Id int `json:"id" gorm:"primary_key"`
-	Account_id int `json:"account_id"`
-	Amount int `json:"amount"`
+	Id int64 `json:"id" gorm:"primary_key"`
+	Account_id int64 `json:"account_id"`
+	Amount int64 `json:"amount"`
 }
